requirements.txt
Django
djangorestframework
gunicorn
